+++ webknossos/webknossos/client/_download_dataset.py
@@ -23,7 +23,7 @@ _DOWNLOAD_CHUNK_SIZE = Vec3Int(512, 512, 512)
 
 def download_dataset(
     dataset_name: str,
-    organization_name: Optional[str] = None,
+    organization_id: Optional[str] = None,
     bbox: Optional[BoundingBox] = None,
     layers: Optional[List[str]] = None,
     mags: Optional[List[Mag]] = None,
@@ -33,11 +33,11 @@ def download_dataset(
     client = _get_generated_client()
     context = _get_context()
 
-    if organization_name is None:
-        organization_name = context.organization
+    if organization_id is None:
+        organization_id = context.organization_id
 
     dataset_info_response = dataset_info.sync_detailed(
-        organization_name=organization_name,
+        organization_name=organization_id,
         data_set_name=dataset_name,
         client=client,
     )
@@ -108,7 +108,7 @@ def download_dataset(
             ):
                 chunk_in_mag = chunk.in_mag(mag)
                 response = dataset_download.sync_detailed(
-                    organization_name=organization_name,
+                    organization_name=organization_id,
                     data_set_name=dataset_name,
                     data_layer_name=layer_name,
                     resolution=mag.max_dim_log2,