+++ wkcuber/format.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+set -eEuo pipefail
+
+if [ $# -eq 1 ] && [ "$1" = "check" ]; then
+    poetry run black --check .
+else
+    poetry run black .
+fi